+++ packages/patternfly-4/react-charts/src/components/ChartContainer/ChartContainer.test.tsx
@@ -1,0 +1,17 @@
+import * as React from 'react';
+import { shallow } from 'enzyme';
+import { ChartLegend } from '../ChartLegend/ChartLegend';
+
+Object.values([true, false]).forEach(isRead => {
+  test(`Chart`, () => {
+    const view = shallow(<ChartLegend />);
+    expect(view).toMatchSnapshot();
+  });
+});
+
+test('renders container via ChartLegend', () => {
+  const view = shallow(
+    <ChartLegend data={[{ name: 'Cats' }, { name: 'Dogs' }]} title="Average number of pets" height={50} width={200} />
+  );
+  expect(view).toMatchSnapshot();
+});