+++ packages/patternfly-4/react-charts/src/components/ChartTheme/index.ts
@@ -1,0 +1,10 @@
+export {
+  ChartBaseTheme,
+  ChartDonutTheme,
+  ChartDonutThresholdDynamicTheme,
+  ChartDonutThresholdStaticTheme,
+  ChartDonutUtilizationDynamicTheme,
+  ChartDonutUtilizationStaticTheme,
+  ChartThemeColor,
+  ChartThemeVariant
+} from './ChartTheme';