+++ packages/patternfly-4/react-charts/src/components/ChartTheme/themes/donut-theme.ts
@@ -1,0 +1,23 @@
+// TODO Replace label props with PF css when available
+
+// Donut styles
+export const DonutStyles = {
+  label: {
+    subTitle: {
+      fill: '#bbb',
+      fontSize: 16
+    },
+    title: {
+      fontSize: 26
+    }
+  }
+};
+
+// Donut theme
+export const DonutTheme = {
+  pie: {
+    height: 230,
+    padAngle: 1,
+    width: 230
+  }
+};