+++ packages/patternfly-4/react-charts/src/components/ChartTheme/themes/donut-utilization-theme.ts
@@ -1,0 +1,39 @@
+// TODO Replace label props with PF css when available
+
+// Donut utilization styles
+export const DonutUtilizationStyles = {
+  label: {
+    subTitle: {
+      fill: '#bbb',
+      fontSize: 16
+    },
+    title: {
+      fontSize: 26
+    }
+  },
+  legend: {
+    orientation: 'vertical'
+  },
+  thresholds: {
+    colorScale: ['#F0AB00', '#C9190B']
+  }
+};
+
+// Donut utilization dynamic theme
+export const DonutUtilizationDynamicTheme = {
+  pie: {
+    height: 230,
+    padAngle: 1,
+    width: 230
+  }
+};
+
+// Donut utilization static theme
+export const DonutUtilizationStaticTheme = {
+  legend: {
+    colorScale: ['#EDEDED', '#D2D2D2', '#BBBBBB']
+  },
+  pie: {
+    colorScale: ['#EDEDED']
+  }
+};