+++ packages/patternfly-4/react-core/src/components/OptionsMenu/OptionsMenu.test.js
@@ -1,12 +1,11 @@
 import React from 'react';
 import { mount } from 'enzyme';
-import OptionsMenu from './OptionsMenu';
-import OptionsMenuToggle from './OptionsMenuToggle';
-import OptionsMenuItemGroup from './OptionsMenuItemGroup';
-import OptionsMenuItem from './OptionsMenuItem';
-import OptionsMenuSeparator from './OptionsMenuSeparator';
-import { OptionsMenuDirection, OptionsMenuPosition } from './optionsMenuConstants';
-import OptionsMenuToggleWithText from './OptionsMenuToggleWithText';
+import { OptionsMenu, OptionsMenuDirection, OptionsMenuPosition } from './OptionsMenu';
+import { OptionsMenuToggle } from './OptionsMenuToggle';
+import { OptionsMenuItemGroup } from './OptionsMenuItemGroup';
+import { OptionsMenuItem } from './OptionsMenuItem';
+import { OptionsMenuSeparator } from './OptionsMenuSeparator';
+import { OptionsMenuToggleWithText } from './OptionsMenuToggleWithText';
 
 const menuItems = [
   <OptionsMenuItemGroup key="first group">