+++ packages/patternfly-4/react-core/src/components/OptionsMenu/optionsMenuConstants.d.ts
@@ -1,9 +1,0 @@
-export const OptionsMenuPosition: {
-  right: 'right';
-  left: 'left';
-};
-
-export const OptionsMenuDirection: {
-  up: 'up';
-  down: 'down';
-};