+++ packages/patternfly-4/react-core/src/declarations.d.ts
@@ -1,7 +1,6 @@
-// We use a JS babel plugin to resolve these imports that `tsc` doesn't
-// know about, so pretend that everything will work
-declare module '@patternfly/react-styles/css/*' {
-  const styles: any;
-  export default styles;
+// This is for charts for jest.
+declare module 'victory-core' {
+  export const Data: any;
+  export const Helpers: any;
+  export const Path: any;
 }
-