+++ packages/patternfly-4/react-table/src/components/Table/BodyWrapper.js
@@ -4,20 +4,24 @@ import styles from '@patternfly/react-styles/css/components/Table/table';
 import { css } from '@patternfly/react-styles';
 import { mapOpenedRows } from './utils/headerUtils';
 
-const BodyWrapper = ({ children, mappedRows, ...props }) => {
-  if (mappedRows.some(row => row.hasOwnProperty('parent'))) {
-    return (
-      <React.Fragment>
-        {mapOpenedRows(mappedRows, children).map((oneRow, key) => (
-          <tbody {...props} className={css(oneRow.isOpen && styles.modifiers.expanded)} key={`tbody-${key}`}>
-            {oneRow.rows}
-          </tbody>
-        ))}
-      </React.Fragment>
-    );
+// eslint-disable-next-line react/prefer-stateless-function
+class BodyWrapper extends React.Component {
+  render() {
+    const { children, mappedRows, ...props } = this.props;
+    if (mappedRows.some(row => row.hasOwnProperty('parent'))) {
+      return (
+        <React.Fragment>
+          {mapOpenedRows(mappedRows, children).map((oneRow, key) => (
+            <tbody {...props} className={css(oneRow.isOpen && styles.modifiers.expanded)} key={`tbody-${key}`}>
+              {oneRow.rows}
+            </tbody>
+          ))}
+        </React.Fragment>
+      );
+    }
+    return <tbody {...props}>{children}</tbody>;
   }
-  return <tbody {...props}>{children}</tbody>;
-};
+}
 
 BodyWrapper.propTypes = {
   children: PropTypes.node,