+++ packages/patternfly-4/react-tokens/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 2.5.3 (2019-05-29)
+
+**Note:** Version bump only for package @patternfly/react-tokens
+
+
+
+
+
 ## 2.5.2 (2019-05-07)